requirements.txt
alembic==1.16.2
api-tool @ git+https://github.com/andreynetrebin/api-tool.git@844919d77050011e87e58d71be36d8962db84ccc
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
Flask==2.1.2
Flask-Login==0.6.3
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysqlclient==2.2.7
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==2.1.2
